Remove "reference" section
I found myself the reference section doesn't really mean "reference", it is just some nice to read article when I develop this document guide. It should be name as "resource", it was added many years ago and I can't remember why I need it as the reference. Therefore removing the entire section, it actually doesn't have reference purpose in this guide.